cape: add pydantic model for v2.2
